requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
et_xmlfile==2.0.0
fastapi==0.127.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
openpyxl==3.1.5
packaging==25.0
paho-mqtt==2.1.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.21
requests==2.32.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
